пункт 3, перенёс весь код из ProfileViewController.swift в новый наследник UIView
